Subiu

+++ src/main/java/br/edu/ifpr/biblioteca/controledebiblioteca/controllers/BookExcluirController.java
@@ -1,16 +1,16 @@
 package br.edu.ifpr.biblioteca.controledebiblioteca.controllers;
 
+import br.edu.ifpr.biblioteca.controledebiblioteca.domain.enums.BookStatus;
+import br.edu.ifpr.biblioteca.controledebiblioteca.domain.models.Book;
 import br.edu.ifpr.biblioteca.controledebiblioteca.repositories.BookRepository;
 import br.edu.ifpr.biblioteca.controledebiblioteca.services.BookService;
-import jakarta.servlet.RequestDispatcher;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "BookExcluirController", value = {"/excluir"})
+@WebServlet(name = "BookExcluirController", value = {"/BookExcluirController"})
 public class BookExcluirController extends HttpServlet {
 
     BookService bookService;
@@ -21,9 +21,17 @@ public class BookExcluirController extends HttpServlet {
 
         String id = request.getParameter("id");
 
-        bookRepository.removeById(Integer.parseInt(id));
+        Book book = bookRepository.findById(Integer.parseInt(id));
 
-        response.sendRedirect("http://localhost:8080/biblioteca/");
+        if(book.getStatus().equals(BookStatus.INDISPONIVEL)) {
+
+            bookRepository.removeById(Integer.parseInt(id));
+
+            response.sendRedirect("http://localhost:8080//ControleDeBiblioteca/BookListarController");
+        } else {
+            response.sendRedirect("http://localhost:8080//ControleDeBiblioteca/BookListarController");
+
+        }
 
     }
 }
